wip options menu - working Audio menu

+++ Assets/Scripts/Models/OptionsObject.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [System.Serializable]
 public class OptionsObject
 {
-    public float musicVolume = 10;
-    public float sfxVolume = 10;
+    public int musicVolume = 10;
+    public int sfxVolume = 10;
 
     public int targetFPS = Options.THIRTY_FPS;
     public bool vSync = false;
